+++ jvm/src/test/scala/com/nawforce/apexlink/pkg/DefinitionProviderTest.scala
@@ -548,16 +548,22 @@ class DefinitionProviderTest extends AnyFunSuite with TestHelper {
   test("SObject variable typename") {
     val contentAndCursorPos =
       withCursor(s"public class Bar { { Foo${CURSOR}__c foo; } }")
+    val objectMetadata = customObject("Foo", Seq())
 
     FileSystemHelper.run(
       Map(
         "Bar.cls"                               -> contentAndCursorPos._1,
-        "objects/Foo__c/Foo__c.object-meta.xml" -> customObject("Foo", Seq())
+        "objects/Foo__c/Foo__c.object-meta.xml" -> objectMetadata
       )
     ) { root: PathLike =>
       val org = createHappyOrg(root)
-      val definition = org.unmanaged
+      val rawDefinition = org.unmanaged
         .getDefinition(root.join("Bar.cls"), line = 1, offset = contentAndCursorPos._2, None)
+      assert(rawDefinition.length == 1)
+      assert(rawDefinition.head.target == com.nawforce.pkgforce.path.Location.all)
+      assert(rawDefinition.head.targetSelection == com.nawforce.pkgforce.path.Location.all)
+
+      val definition = rawDefinition
         .map(LocationLinkString(root, contentAndCursorPos._1, _))
 
       assert(
@@ -566,8 +572,8 @@ class DefinitionProviderTest extends AnyFunSuite with TestHelper {
             LocationLinkString(
               "Foo__c",
               path"/objects/Foo__c/Foo__c.object-meta.xml",
-              "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n",
-              "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
+              objectMetadata,
+              objectMetadata
             )
           )
       )